Make chevron component more resilient
The chevron component was developed quickly and under time pressure, and therefore was quite brittle to any changing content within it as we hadn't expected the text to change. This rewrites the component to handle varying text lengths better, and remove some of the hacks we had to put in place to get it to work.