Simplify markup when having only one document
When using the document list component in conjunction with a table we end up with a list of one item in each cell which makes it harder to navigate using assistive tech. This updates the component so that a document is presented without a wrapping list when only one document is provided.