Delete unconsented cookies before explicit preferences set
At the moment, if a user had visited GOVUK recently and had cookies set, and they also visited yesterday and got the new default consent cookie, all their existing cookies would remain and not be deleted. This is because the cookie deletion wasn't quite working and cookies are only deleted when setting the consent cookie. The only way for this person to delete their cookies would be for them to 1) manually delete them or 2) for them to set their preferences on the settings page.

This commit adds a method to delete all unconsented cookies when the cookie banner is present, so people in the above position will have previously set cookies deleted for them rather than having to manually delete them.