Remove ordered related items from breadcrumb logic
The ordered related items breadcrumbs logic relies on the content item having a parent in the links hash.
This isn't always present (in fact, it seems to be a feature of mainstream browse and HTML attachments rather than ordered_related_items).
This means documents with ordered_related_items have empty breadcrumbs, which isn't ideal, especially when they're tagged to taxons which could be shown.
This is more of a temporary fix/workaround until we make a decision on what we expect the breadcrumb behaviour to be when ordered_related_items are present.