Update modules script to work with any module type
Previously the modules script was dealing only with GOV.UK Frontend Toolkit and GOV.UK Publishing scripts. It now handles GOV.UK Frontend modules initialisation too. This means we don't need to call `GOVUKFrontend.initAll` anymore, but we replace this with an empty function for backwards compatibility (othwerwise it will cause a hard error on projects calling this function)