Check that details component exists
Before we add an event listener to the details component we should check that
the selector has actually found something - or things go wrong.

Unlike `addEventListener`, jQuery's `.on('click', function() { ... })`
silently handles nothing being found. This is because the jQuery selector
_always_ returns an iterable jQuery object.

The `.on()` listener loops through the jQuery object given to it - which means
that if the object has a length of zero then nothing happens and nothing
fails.

Using `querySelector` gives an HTMLElement object if the element is present or
`null` if there are no matches.

Trying to add an event listener to `null` throws an error - this is why a
check is required to see if the element exists before trying to add a
listener.