Mobile search toggle button
The button that opens/closes the layout header search form on mobile is
not actually a button, but an anchor tag. Therefore, it does not behave
as a button (for instance, it does not activate when using the space
key)

This is semantically incorrect and cannot be easily understood by screen
reader users, or used by voice control users.

Use button markup instead of an anchor tag, to more closely match the
presentation and function of this element.

Hide element when Javascript is disabled as the element is not
functional without JS.