Update layout header component
Add data attribute support on the layout header navigation links. The
reason for this is because we are intending to use the layout header
component for the transition checker nav whose links have data
attributes for tracking.
Similarly, add support for link IDs. This is for the Slimmer logic that
hides/shows different links depending on the user's logged in status.