Fix layout_header CSS issues
Currently, the layout header component looks different in the context of
static vs in isolation.
- on the account manager, which does not use static, the layout header
navigation does not align as intended with the rest of the header on
tablet devices
- in the context of an application that uses static, the layout header
  inherits some styles from static. Additionally, application CSS will
conflict with the layout_header CSS resulting in spacing issues.

This aims to address some of these problems by removing unnecessary
classes, as well as increasing layout_header CSS specificity where this
CSS might be overruled by external styles from either static or the
parent application's stylesheet.